Overhauled configuration and implemented default track system
The configuration has been changed from a ugly beast riddled with many
<NEWLINES> and replaced with a system that takes advantage of many YAML
features that Karma did not support previously, which made the underlying
plugin code for different features much easier and cleaner.

There is now a `default' configuration option for tracks which specifies
if the track is the default track for new players or not.