Allow changing the track a player is in.
Adds a command: /k track. This command allows you to change the track that
a player resides in. It will automatically increase the player's Karma
point value to that of the track's start group, if necessary. It will
switch the player's group into the one that his current Karma point value
makes him eligible for.