Fixes Bug 3 - Process demotetofirstgroup in automatic mode.
This issue was caused by Karma processing this option when ranks were
being changed instead of when points were being modified. This caused
points and the player's rank go out of sync.

The problem was solved by moving the process from the permission updater
into the automatic remove points method.