Fix unit tests for java 7 powermockito
After updating the signature of methods in the database classes, I
knew I would have to make minimal changes to the unit tests to make
them continue to work properly. I did not expect that PowerMockito
would throw a fit and break everything. Apparently updating the
project to java 7 and my computer to java 8 did not suit the tests
well as the PowerMockito version I depended on did not support
either of those. I also had to move an interior class to package
level because of some wierd errors from PowerMock. It might be safe
to move it back into the class because I updated PowerMock, but since
the tests are working properly, i'm not going to mess with it.